Seems to work reliably now -- don't touch it
Previously the radio message responses were often, but not always, lost. This
was the case even if the radio was always listening. What appears to have fixed
it: returning on unhandled button events without calling
twr_scheduler_plan_now(0). Hypothesis: spamming the scheduler with heavy tasks
like LCD updates can deny the radio handler from running in time, causing it to
lose the response to the message that was just sent.

Use the encoder module wheel to control the UI. The encoder wheel is a much
nicer UI than the two LCD buttons. Unfortunately it is difficult to use both
interfaces at the same time. The encoder button is connected to BOOT. Both LCD
buttons are also connected to BOOT as well as separate inputs, so if button
handlers are set up for all three button inputs, the LCD buttons will generate
twice the events.

Additionally, some minor bug fixes and compiler warning fixes.